OPS-68: CLean up database connections so we can move this to production server
